+++ CHANGELOG.md
@@ -129,6 +129,7 @@
 - [10393](https://github.com/vegaprotocol/vega/issues/10393) - Fixed source weight validation.
 - [10396](https://github.com/vegaprotocol/vega/issues/10396) - `ListTransfers` returns error.
 - [10299](https://github.com/vegaprotocol/vega/issues/10299) - `ListTransfers` does not return staking rewards.
+- [10378](https://github.com/vegaprotocol/vega/issues/10378) - Ensure network position has price set at all times.
 
 ## 0.73.0